pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dad-can-i-wear"
dynamic = ["version"]
description = "Weather-based clothing recommendations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.2",
    "httpx>=0.28.1",
    "jinja2==3.1.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "uvicorn==0.27.1",
]

[tool.hatch.version]
path = "app/__about__.py"

[tool.hatch.build.targets.wheel]
packages = ["app"]
